+++ pytsite/content/_widget.py
@@ -2,7 +2,7 @@
 """
 from pytsite import taxonomy as _taxonomy, auth as _auth, widget as _widget, html as _html, lang as _lang, \
     router as _router, tpl as _tpl, odm as _odm
-from . import _model, _functions
+from . import _model, _api
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -17,7 +17,7 @@ class ModelSelect(_widget.select.Select):
 
         items = []
         u = _auth.get_current_user()
-        for k, v in _functions.get_models().items():
+        for k, v in _api.get_models().items():
             if self._check_perms:
                 if u.has_permission('pytsite.odm_ui.browse.' + k) or u.has_permission('pytsite.odm_ui.browse_own.' + k):
                     items.append((k, _lang.t(v[1])))
@@ -48,7 +48,7 @@ class EntitySelect(_widget.select.Select2):
         # In AJAX-mode Select2 doesn't contain any items,
         # but if we have selected item, it is necessary to append it
         if self._ajax_url and self._value:
-            self._items.append((self._value, _odm.get_by_ref(self._value).title))
+            self._items.append((self._value, _odm.get_by_ref(self._value).f_get('title')))
 
         return super().get_html_em()